Fix printing on Windows when sandbox is enabled
Auditors:  @bridiver

This makes it closer to how it works in Chromium and fixes the forever spooling problem on Windows.

The old way worked when the sandbox was disabled but failed on the call for shared_buf.CreateAndMapAnonymous when the sandbox was enabled

This also fixes a handle leak which Chromium already had fixed.