Install enabled by default
We were sending an async message to the extension system to
install the extension, and then calling the callback sync which would
enable.  But enable wouldn't find the extension because it wasn't loaded
yet.

This fixes that race condition as reported in #4425.

I know we may want to load without enabling at some point,
but we don't do that today, so I didn't bother adding a different event
for that yet.

There is also a browser-laptop issue coming soon.

Auditors: @bridiver