Re-add deprecated screen module to remote
Auditors: @darkdh

It should be deprecated and is only used in tests. When we eventually remove it we can just store the value in appState from the main process and use it in tests.

But for now it's easier to just add it to the deprecation list instead of having the sudden removal.