URLRequestContextGetter: Do not set the URLRequestContext for NSS
It was wrong because it's set everytime we were creating an URLRequestContext.
It must be done only once. It was moved to the recently added IOThread which
uses the system URLRequestContextGetter for NSS.

In order not to pull in too many dependencies we provide a simplified version
of DefaultNetworkContextParams which does not use the policy service to
retrieve the http_09_on_non_default_ports_enabled setting.