Put the control auth cookie in the watch directory too.
This is necessary because there is an unfortunate ordering issue with
tor startup: it writes the control port first, and then the auth
cookie, but we need both in order to connect to the control port.
And it doesn't delete the auth cookie, so it can get stale.  Hence we
need to monitor writes to the auth cookie too.