Issue 14968: NULL deref while setting the window_id for the tab_helper of new_contents
For the source webcontents owner_window or relay will not be available if the tab is not attached. Instead we fetch the browser directly from the tab_helper to get window_id for the new_contents.

For extensions and main window the tab_helper will be null, if tab_helper is null we fallback to the original code.

Auditors - @bsclifton, @brianjohnson, @darkdh